separate main code into multiple components
